requirements.txt
Faker==18.9.0
kafka-python>=2.0.2
psycopg2-binary>=2.9.3
pytz==2023.3 